ENH: avoid copy on sparse __init__ when dtype is given
Sparse matrix constructors are supposed to avoid copies when copy=False
is given. However, when dtype is given explicitly, a copy is made even
if not needed.

Also when copy=True is given, explicit dtype will force two copies
instead of one.

This commit avoids the extra copy and adds generic tests for these
cases.

+++ scipy/sparse/bsr.py
@@ -209,7 +209,7 @@ class bsr_matrix(_cs_matrix, _minmax_mixin):
                 self._shape = check_shape(shape)
 
         if dtype is not None:
-            self.data = self.data.astype(dtype)
+            self.data = self.data.astype(dtype, copy=False)
 
         self.check_format(full_check=False)
